+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Added Manual Changelog
 * Changed Latex Processing Unit to make crafted dry rubber instead of tiny ones with increased fluid latex req
+* Changed Latex Processing Unit recipe to not use Latex Bucket