+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Added Manual Changelog
 * Changed Latex Processing Unit to make crafted dry rubber instead of tiny ones with increased fluid latex req
 * Changed Latex Processing Unit recipe to not use Latex Bucket
+* Fixed Bioreactor efficiency bar not resetting and buffed biofuel production rates